Feat(dot/parachain): Add `CompactStatement` type (#4424)
Introduced the `CompactStatement` and `compactStatementInner` types. We will only deal with `CompactStatement` in other codebases. `compactStatementInner` is only used for encoding/decoding of `CompactStatement`, which is why it's not exported.

Implemented `MarshalSCALE` and `UnmarshalSCALE` methods to have custom encoding/decoding logic for `CompactStatement`.

Also, the sign logic of statementVDT has been updated to encode the data properly before signing.